+++ parser/src/test/java/io/github/cowwoc/styler/parser/test/TokenTest.java
@@ -1,7 +1,7 @@
 package io.github.cowwoc.styler.parser.test;
 
-import io.github.cowwoc.styler.parser.*;
-
+import io.github.cowwoc.styler.parser.Token;
+import io.github.cowwoc.styler.parser.TokenType;
 import org.testng.annotations.Test;
 
 import static io.github.cowwoc.requirements12.java.DefaultJavaValidators.requireThat;